New blspy, key formats, and new plot format (#309)
* Bump blspy-ietf
* wip lots of errors
* Respond to bram feedback, use eip 2334 for keys
* Fix most tests, use latest blspy
* Improve simulation test constants
* Fixed issue with plotting and reading plots
* Fix simulation test
* Override puzzle hash if we don't have key, for safety. Bump protocol version
* Fix superlint issue
* Switch to 1.0 proof of space format
* Temporary genesis block
* Update chiabip158, chiavdf versions

+++ tests/test_spendbundle.py
@@ -8,12 +8,13 @@ from src.util.condition_tools import (
 from src.wallet.puzzles import p2_delegated_puzzle
 from src.wallet.puzzles.puzzle_utils import make_create_coin_condition
 from tests.keys import puzzle_program_for_index
+from src.util.ints import uint32
 
 
 def test_1():
-    puzzle_program_0 = puzzle_program_for_index(0)
-    puzzle_program_1 = puzzle_program_for_index(1)
-    puzzle_program_2 = puzzle_program_for_index(2)
+    puzzle_program_0 = puzzle_program_for_index(uint32(0))
+    puzzle_program_1 = puzzle_program_for_index(uint32(1))
+    puzzle_program_2 = puzzle_program_for_index(uint32(2))
 
     conditions = [
         make_create_coin_condition(std_hash(bytes(pp)), amount)
